Print the Nth Markov Number
This will print the nth markov number starting from index 0.  I tested this for up to N=39, which is as far as they go on the OEIS for this sequence. 
After that I am not sure, to get an accurate result you may need to increase the search term, n**3 to a higher power.