Add original implementation
The previous implementation used the library function, which does not provide value to the repository. 

This implementation returns the nth element from the positive indexed Euler numbers. This is OEIS A000364. I chose this sequence because the original sequence has every odd entry as 0. It also has alternating negative entries. Implementing these wouldn't serve much purpose. 

It works off of the definition that the nth Euler number is the number of downup permutations of [2n].